refactor: delete dead code in src/tools/playbooks/
Inline playbook tool implementations (create, list, update, delete) directly
into bundled-skills/playbooks/tools/ and update test imports. The old
src/tools/playbooks/ directory contained unregistered handler stubs that
only served as an unnecessary indirection layer.
