+++ assistant/src/__tests__/playbook-execution.test.ts
@@ -41,12 +41,10 @@ import { memoryItems } from '../memory/schema.js';
 import { computeMemoryFingerprint } from '../memory/fingerprint.js';
 import { v4 as uuid } from 'uuid';
 import type { ToolContext } from '../tools/types.js';
-import {
-  executePlaybookCreate,
-  executePlaybookList,
-  executePlaybookUpdate,
-  executePlaybookDelete,
-} from '../tools/playbooks/index.js';
+import { executePlaybookCreate } from '../config/bundled-skills/playbooks/tools/playbook-create.js';
+import { executePlaybookList } from '../config/bundled-skills/playbooks/tools/playbook-list.js';
+import { executePlaybookUpdate } from '../config/bundled-skills/playbooks/tools/playbook-update.js';
+import { executePlaybookDelete } from '../config/bundled-skills/playbooks/tools/playbook-delete.js';
 
 initializeDb();
