--- assistant/src/tools/playbooks/index.ts
@@ -1,4 +1,0 @@
-export { executePlaybookCreate } from './playbook-create.js';
-export { executePlaybookList } from './playbook-list.js';
-export { executePlaybookUpdate } from './playbook-update.js';
-export { executePlaybookDelete } from './playbook-delete.js';
